Fixed issues with Razor compilation
The razor compilation now works but basically forces the user to use an
ExpandoObject as the model. I would prefer it being a straight
dynamic/use whatever but I can't seem to get that to work with
RazorEngine.  Moved the email stuff to the appropriate directories.
Started work on adding helper functions to the controller base class.
Added more logging information as well.