Fixed issue with image embedding with CSS files
I forgot that to add in the other file types for images (jpeg, gif, bmp,
etc). This has now been fixed and it should reduce the final CSS files'
size. Also finally where automated deployment is possible (still need to
set up Jenkins and fix the default package, but other than that we're
good to go).